move all of the bundles for getting the location info into the /location-bundles folder
