+++ src/app-pages/map/components/map-details/components/time-series/TimeSeriesDateRange.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "redux-bundler-react";
-import { radarTimeControls } from "../../../../../../app-bundles/radar-time-series-bundle";
+import { radarTimeControls } from "../../../../../../app-bundles/location-bundles/radar-time-series-bundle";
 import { dateToString } from "../../../../../../utils";
 
 export const TimeSeriesDateRange = ({
